requirements.txt
streamlit 
Pillow 
torch 
torchvision 
matplotlib
opencv-python
joblib
pymorphy3
scikit-learn
umap-learn
nltk